Extracting functions from Main class and adding some protocols to improve the segregation
